fix dial_fuel
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@517 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/PlayerEntity.m
@@ -2383,9 +2383,9 @@ double scoopSoundPlayTime = 0.0;
 {
 	if (fuel <= 0.0)
 		return 0.0;
-	if (fuel > PLAYER_MAX_FUEL)
+	if (fuel > (float)PLAYER_MAX_FUEL)
 		return 1.0;
-	return fuel / PLAYER_MAX_FUEL;
+	return (double)fuel / (double)PLAYER_MAX_FUEL;
 }
 - (double) dial_hyper_range
 {
